[#139] Improvement: externalize Gravitino metadata and Jupyter notebooks outside of containers (#140)
### What changes were proposed in this pull request?

Mount a local folder to the Gravitino and Jupyter container, which
persistents the metadata of Gravitino (h2 db file by default) and the
notebooks of Jupyter. So that they won't be lost when the containers be
re-created.

### Why are the changes needed?

No changes for other part. 

If user do want to cleanup the metadata, they need manually clean up the
"data/gravitino" or "data/jupyter" folder under the playground folder.

Fix: #139

### Does this PR introduce _any_ user-facing change?

No

### How was this patch tested?
I tested it in my local. When I stop, drop the container, and then
restart, previous metadata of Gravitino and Jupyter were kept.